project/requirements.txt
ipykernel
openai==1.14.0
jupyterlab
python-dotenv
yfinance
pandas
matplotlib